Dynamic programming, hashing, sorting of complex subformulas, reversed order of sorting of atomic subformulas, private ofsf_subf, changed verbose output, switch rlqeheu if off by default now. Container access functions have been renamed; hopefully also pasf is completely adapted to this. Fixed a bug in pasf; where lessp and minus was used on domain elements.
Due to the new sorting of formulas the rlg files changed considerably. Some
tests run longer because there are different terms chosen for Gauss; in
particular, rp2 yields a larger result. clprl.tst had to be adapted. lpdo.tst
yiels a different sample factorization for one example. The support for sorting
in qqe_ofsf was incomplete and incompatible with the code changes in ofsf.

New switches: rlsisocx (on), rlqeidentify (off), rlqedyn (on), rlqesubf (on),
rlqevb (on), rlqevbold (off).


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@1713 2bfe0521-f11c-4a00-b80e-6202646ff360