This continues review of the arithmetic code in CSL. I am part way through looking at all shifts to try to ensure that I remain within the C++ standard and so not rely on undefined features such as the effect of left shifts on signed types. If there is a 128-bit intgeger type available I exploit it in the hope that that speeds up multiplication of large fixnums, division of numbers of length up to 120 buts biy one of length up to 60 bits, and modular multiplication for moduli up to 2^59.
Meanwhile a glich that was little more than a typo has led to the FOX GUID
leaving the menu entries for some switches greayed out. I believe this
checkin should fix that. 

The configure scripts for CSL now use a useful autoconf extension that lets
then arrange that compiler flags are set to make certain that the C++ compiler
used supports a reasonable fraction of C++-11. If C++-14 support seems to
be available then that is enabled.


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@3838 2bfe0521-f11c-4a00-b80e-6202646ff360