Another round of CSL internal changes! These should mend a couiple of glitches in tracing, but are mostly preparation for the next step, which ought to get trace output for interpreted code able to report what function something was called for and should also be able to reinstate traceset. This bit of re-factoring is widespread enough that it needs to be its own check-in.
I also arrange that going just "make" at the top-level should observe when
you have several variants of CSL configured and build them all in parallel.
This is based on a hypothesis that few people have HUGE numbers of versions
configured in the same tree and almost everybody now has a computer with
a multi-core CPU and (say) at least a gigabyte of memory, so that the
parallel build is feasible and shoudl complete somewhat faster.


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@3901 2bfe0521-f11c-4a00-b80e-6202646ff360