This is a first step towards better floating point handling in PSL, first for 64bit Linux.
The goal is for the compiler to generate floating point instructions directly instead of
jumping to C code for floating point artihmetic. Apart from the performance penalty,
exceptions raised in C code are a pain to handle from the Lisp side.

This commit introduces a new fluid fp-except-mode*. If set to 0, floating point operations
do not cause errors but return the appropriate non-finite value (NaN, +/-Inf). For all 
other values of fp-except-mode*, error are raised. The default value is currently 1.

Note: This interface for controlling floating point behaviour may change in the future.
In particular, setting the variable fp-except-mode* may be replaced by calling a function.


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@3933 2bfe0521-f11c-4a00-b80e-6202646ff360