More work on rationalization of the code for plus and difference so that the dispatch on all the various arithmetic types is better implemented. Code in csl/cslbase/bigarith.red tries to test integer arithmetic. That is designed for us with CSL but I have tried retrofitting some PSL support, but that is not (yet) fully finished or debugged. I will still want to go through and re-do the type disptch for times and the division/remainder functions, and also the smaller number of unary generic arithmetic functions. But maybe bigarith.red will help me to find both long-standing and any newer arithmetic bugs. I will then need a similar style file that will check arithmetic with floating point components, but I will write that assuming that integer arithmetic is OK.
I have changed rprint so it can cope with "implied progn" in COND and at the
top-level of procedure definitions. I have improved the "make foo.cxx" rule
in Makefiles to allow for versions of sed, and put in "make foo.s" so you can
inspect the generated assembly code for things if you are the sort of
person really fussy about absolute performance.


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@3958 2bfe0521-f11c-4a00-b80e-6202646ff360