The bulk of the factorizer code was written between 1979 and 1981, ie not far short of 40 years ago. The core ideas are notionally straightforward but the Zippel-style treatment for sparse polynomials ends up with a great many intertwined tests and many special cases. The basic idea is that while factorizing (or GCDing) you form images of input polynomials by substituting small integers for all but one of the variables and then pick a prime and reduce the coefficients in an eventual univariate image modulo that prime. Then you reconstruct, with the Hensel Construction the main tool. But when done in a straightforward way the Hensel Construction has very bad consequences when your polynomials are sparse and the evaluations you have made do not give a faithful image of true results - you tend to reconstuct a dense polynomial and old stop when you hit a coefficient or degree bound (or more probably when you run out of time and space). The spareseness preserving schemes use an heuristic that if some coefficient is zero in an image polynomial perhaps it really is zero in the full result, so you use a variant on reconstruction that assumes that and sees if it leads to a valid full result. But of course especially if you are using small integer value substitutions and for awkward cases you can get zeros in an image that are a consequence of cancellation in the image-forming step and so the path taken will not work. That situation may not become apparent until a significant amount of work on sparse interpolation has been done. It can (and does) arise because of multiple special cases all arising together. In the checkin today I have disabled one of the bits of special case detection that presumably saves resources when it works but that was leading to trouble in other cases. I have disabled it as an admission of my inability to work though all the ancient code and recover all the reasoning behind it (even though "on trfac" and "on trallfac" still produce what I count as being remarkably carefull trace output). This chance may be presumed to slow factorization and GCD down in some cases, but it also fixes a failure in others. I would really welcome anybody who had the time and energy to review all that code rather carefully and check the logic carefully both for special cases currently not exploited and for paths that could lead to misery.
The second check-in is in poly/polrep.red where OLD code commented out some
years ago worked by temporarily resetting a global flag and restoring it
after some critical code. That had been replaced with a version that rebound
the global flag so that the variable was automatically restored whether the
inner code exited cleanly or via an error. I have just removed the old
commented ouut (and dangerous) code. The reason I note that here is that the
factorizer as whole has a number of places where it saves and restores
global state in the older active-code manner, so if it exits because of
an abrupt error the global state may be left in a state that is not as it
was on entry. That could of course have a range of "interesting" and
potentially delayed side effects and break things that were apparently remote!
I know what some of the cases are but would not pretend to have spent the
time needed to make an audit of all of them or to assess just how much
damage could arise, but that does mean that continuing any calculation after
the factorizer (and in fact plenty of other parts of Reduce!) has reported
an internal error requires very careful thought it if it is to lead to
reliable recovery.


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@3981 2bfe0521-f11c-4a00-b80e-6202646ff360