Several headings here: (1) the test for an integer being prime now uses the best hashed method I have found so far for 32-bit numbers. Then a deterministic test that I can optimise later for values up to 64-bits. After that a BPSW test (Miller-Rabin followed by Strong Lucas) where there are at present no known false positives at all, and plausible estimates for the first such are amazingly high.
(2) In a bunch of places where there were #if statements testing for CSL or
PSL I have adjusted the tests so that they check for the feature involved.
An effect of that is that when or if a Lisp system implements the function
then Reduce will automatically take advantage.

(3) I have put in come comments to explain that having an inline function
that references global or fluid variables is a bad thing, because the
declarations exist within the source file where the inline function happens to
be defined, but the code is then processed potentially somewhere quite
different when it is expanded inline. That explains the parse-time arning
message that such usage leads to.

(4) I have tried a change to what I might describe as the "fat binary"
scheme used in windows binaries created by the installer so that the
temporary files created there are more likely to be cleared up in the case
that Reduce is cancelled or crashes. I have NOT tested that yet - I can most
readily do so once the change is checked in.

(5) the functions MSD and LSD are now more consisten with one another. This
represents an incompatible change in LSD, which used to be more like "count
trailing zeros" than "tell me where the lowest set bit is". This changes the
value returned by 1.

(6) Comments inserted near the definition of ordp to highlight the way that
there are 3 verions in the code and that leading packages can cause switches
between them in ways that could count as confusing and bad. I want this to
trigger a code review that ends up with there being just one version of that
function!


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@4088 2bfe0521-f11c-4a00-b80e-6202646ff360