This is a a first cut towards making "make install" work on at least some platforms. At present it should be viewed as thoroughly experimental, and there is no "make uninstall". So far if you are building on Linux and use "sudo make install" it should try to put file in a structure just like the one set up in the *.deb or *.rom files. If you have used the default setting for "configure" these will go in /usr/local, so you may wish to consider "--prefix=/usr" if you want to put stuff in /usr/bin etc. I have not yet understood all of what I may want to do for Windows and Macintosh. I have do not support the GNU "DESTDIR" option. This is not yet really tested and because it will tend to put files in public places I might suggest that you try it in vitual machines where you have taken a snspshot so you can roll back from it - at least for now. But I hope some of you will try it and provide feedback and comments.
If you use "make install" from the trunk directorry you should be able to say
"make install csl" or "make install psl" and by default it should install
both (if they have been configured). In that case it should install the
version that running bin/redcsl or bin/redpsl would from there.

If you go "make install" in a directory such as cslbuild/<machine>/csl then
just that version is installed, and of course it may be a debugging version or
one configured with other odd options.

At present rfcsl and rfpsl do not get installed by this. I intend to add them
later - to keep what happens with this form of install as similar to that
which happens via the *.deb/*.rpm route as I can. 


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@4117 2bfe0521-f11c-4a00-b80e-6202646ff360