The change in cslbase/isprime.cpp is a further step reducing the size of the tables. cslbase/Makefile.am removes a hack I put in a month or so back that was to smooth a transition awat from using make double-colon rules. It also tried to tidy some messages up a little.
rlisp/proc.red is there to set a place marker. At this stage it does not
activate any changes by default, but it is to signal what I intend to do
and make it easier to test its consequences. If the flag !*acn is set it
alters the parsing of procedure headers so that
(1) Cases like 
      symbolic procedure foo(args:type1, arg2, arg3:type3):resulttype;
      ...
    will be parsed. At present the extra type information is instantly
    discarded and at present type identifiers can only be single words,
    but I view both of those as temporary limitations.
(2) It ceases to be possible to go
      infix newop;
      symbolic procedure u newop v;
      ...
    with the procedure header line written using infix syntax. Various
    other unusual and perhaps related cases may also now be rejected.
    On the other hand I may not have put in as many checks and my new error
    recovery might not be as good.

I believe that the current Reduce sources still compile if !*acn is on. But
I will also want to adjust "decl" so that one can also use type annotations in
declarations of variables at the top of a block and with lambda forms.
The plan is to allow things like
    begin
      scalar save_n:integer:=n, ...
      ...
so that as well as type annotations it will be possible to put initialization
into variable declarations. And then - obviously - to work on developing a type
model that uses the extra information provided.

This relates to two existing chunks of work that are within Reduce. The
"assert" package provides related syntax under "asserted procedure ...".
Because that uses its own procedure header parser what I am doing here should
not hurt things. I wish to concentrate on static type analysis not the dynamic
checking that "assert" does, and that is liable to mean that the way I wish
to generalize type descriptors will end up changing. Also at present "assert"
is not concerned with annotating local variable with types. But I hope it will
be possible to have some cross fertilization. Then "reduce4" introduced syntax
again related to what I am installing now. My ideal will be to get to a state
where the type annotations that I introduce can somewhat take over from the
reduce4 experiment, and if that is ever the case then "reduce4" as such can be
removed as a separate world so there is just "reduce". I am keeping my
code distinct from the !*reduce4 flag because at least to start with I want
to make my extensions compatible with existing Reduce, so my hoped for future
is that test on !*acn go away and the newer code that accepts type annotations
becomes the version generally in use... and that if type annotations are not
present in some segment of code then it will continue to work as at present. 



git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@4133 2bfe0521-f11c-4a00-b80e-6202646ff360