I have re-worked the rlisp parser a bit. It can now allow single words to appear in a place where they might be type specifiers in procedure headings and in the variable declarations at the start of a begin..end block. These are at the moment limited to just arbitrary symbols as type specifiers (I expect to extend on that later), and at present the extra information is not really used. Well for a procedure it ends up stashed on the property list of the procedure name, and for local variable the generated (prog...) block in the parse tree has the type annotations grafted in concealed within a QUOTE where it will not lead to any calculation but where it is potentially available for analysis.
Typical examples (not yet used, and at best for experimentation as yet) might
be
   symbolic procedure addf(u:sf, b:sf):sf; ...
or
   begin scalar a, b:integer_list, c:sq; ... end
and I can also cope with infix procedures as in
   symbolic procedure a:term .+ b:sf : sf; ...

In CSL the procedure type information (and also information about the
file and line-number of definition) is saved in a way such that load!-source()
will restore it to property lists.

Throughout much of Rlisp I have tried a small pass at code tidying. As part of
removing a couple of labels I put support for REPEAT statements into the Lisp-
coded bootstrap parser. I also note that cases that used to read "a eq 'x" can
be rewritten as "a = 'x" because the quoted expression is liable to be a
symbol, and then the compiler can optimise use ofd EQUAL to use of EQ, but the
use of the "=" symbol is probably clearer.



git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@4136 2bfe0521-f11c-4a00-b80e-6202646ff360