A first observation is that I am aware that this checkin breaks the LALR package. But all the other tests seem to run OK at least on Windows so I want to check all this in to make it easier for me to test on other platforms and so that there is a chance of bug reports from others of you.
I have re-worked the way that CSL handles functions with many arguments.
This should be a proper response to a limitation that prevented somebody's
real calculation from completing a month or so ago, and may count as a good
idea anyway. The basic scheme is now that a function call written as
f(a,b,c,d,e,f) with size (say) arguments is internally mapped to
f(a,b,c,list(d,e,f)) so that the first 3 arguments are passed in a native
way and any beyond that as a list. In Reduce as a whole only a small
proportion of all function calls use 4 or more arguments, and my new
implementation in fact should simplify and speed up calls to functions with
0 or 3 arguments compared with older CSL. This is obviously a fairly widespread
change within CSL, but it ought not to be visible at the Reduce level, even
at the symbolic mode Reduce level - unless I have left in bugs!

Whiel doing this I have re-worked the CSL foreign function scheme to use libffi
rather than my own ugly code. This now lets me support foreign function
calls with up to 20 arguments. Right now I have not reinstated support for
calling vararg foreign functions - if anybody is relying on that please bug me!
Also I have implemented but not really tested the new code there.. so again it
will help if anybody who likes foreign functions has a try.

Originally CSL was coded in C. I am gradually migrating it to be more and more
idiomatic C++. But to a real C++ programmer it will still betray the fact that
it is warmed-over C. The most recent change on that front is that the
interface with the FOX library now uses namespaces just a little bit better.
That change means that to build the new version it will be essential to
recompile FOX. I think that ought to happen automatically, but I would still
suggest that "rm -rf cslbuild; ./configure ..." gives the cleanest path
forward with least chance of confusion. Because there is a fair opportunity for
confusion.


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@4188 2bfe0521-f11c-4a00-b80e-6202646ff360