GNU Emacs REDUCE IDE
Fix a bug in reduce-input-file introduced in the previous update, which prevented it correctly switching to a REDUCE process buffer when called interactively.
On Windows (only), use the most likely full paths for the commands to start REDUCE, since the bin directory is not automatically added to the search path.


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@4289 2bfe0521-f11c-4a00-b80e-6202646ff360