sl.el: Standard LISP on Emacs Lisp updates
Added declarations to allow use of the Emacs source code debugger (edebug) on forms defined using macros, in particular prog.

Improved compress, intern, flagp, get, liter.

Redefined prog as an alias for cl-prog (which is not included in the Emacs Common Lisp manual, otherwise I would have used it before) and redefined return as an alias for cl-return.

Redefined errorset as a macro so that the tr (trace) argument can enable an Emacs backtrace provided debug-on-error is true.

Added macro eval-forms-as-sl and command sl-eval-buffer.

This version of sl.el will process input of the form

(begin2)
a_very_simple_form;
end;


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@4291 2bfe0521-f11c-4a00-b80e-6202646ff360