Standard LISP on Emacs Lisp updates
Revised putd and getd (and dm) to handle macros properly (I hope). Standard Lisp macros are different from Emacs Lisp macros: they are nospread and so take a single parameter that gets a list of actual arguments, and they are called with the macro name as the first actual argument. So putd, dm and getd have to map between the two styles. I can now read all of "rlisp.red" except for modules list and switch. I suspect the problem here might be that my definitions of explode and/or compress are still not correct.


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@4333 2bfe0521-f11c-4a00-b80e-6202646ff360