Standard LISP on Emacs Lisp updates
Revise the composites of CAR and CDR (CAAR etc.) to use capitalized versions of the definitions used in Emacs Lisp, which doesn't require much more code.

As a potential step towards better integration with the REDUCE read-eval-print loop, revise READCH optionally to read from the minibuffer as terminal input, and revise my earlier definition of the conventional Standard LISP read-eval-print loop, which reads Emacs Lisp from the minibuffer, also to use the minibuffer for READCH input.

Include an implementation of ORDERP.


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@4390 2bfe0521-f11c-4a00-b80e-6202646ff360