Standard LISP on Emacs Lisp updates
Echo input from a file (as I already do from the minibuffer) and add support for the switch *ECHO to control echoing. As far as I can see, *ECHO is not defined in the Standard LISP report but is assumed by REDUCE. Try harder to force output to the end of the output buffer, since at present it can appear elsewhere if point if moved interactively between inputs.


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@4488 2bfe0521-f11c-4a00-b80e-6202646ff360