Standard LISP on Emacs Lisp updates
(put 'NIL 'QUOTENEWNAM nil) doesn't work because GET can't distinguish between a property that is not set and a property with a value of nil. Hence, NIL is not reliably downcased and I don't want to gamble that NIL will always be evaluated before it's value is used. So instead of using the newnam and quotenewnam mechanisms, I now downcase LAMBDA, NIL and T in COMPRESS.

Changed the debug specs for the AND and OR macros, since trying to use the specs for `and' and `or' doesn't seem to work.

Allow RDS to redirect input to the Emacs Lisp reader, since the bootstrap process needs to be able to read a mix of Lisp and REDUCE input. Tidy up input echoing and ensure that minibuffer input is always echoed.

Provide CONCAT as an alias for `concat' so that I can use it in REDUCE.


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@4492 2bfe0521-f11c-4a00-b80e-6202646ff360