Standard LISP on Emacs Lisp updates
Updates and an initial build system for ESL REDUCE 3.8.

These files should build all of REDUCE 3.8 although I have not yet tested it with more than the main third of the packages. The resulting system runs some test examples but fails others and still needs a lot more work. I think the next step is to improve the debugging support!


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@4526 2bfe0521-f11c-4a00-b80e-6202646ff360