GNU Emacs REDUCE IDE
First check-in in a separate file of a complete re-implementation of the code to highlight matching group and block delimiters, based on the Emacs 26 code to highlight matching parentheses. This needs a lot more testing, but I hope it avoids throwing an error on a mismatched end, as is used to mark the end of a REDUCE source file. This new version supports optionally matching a delimiter when point is immediately inside it, or outside and separated only by white space. It also supports optionally highlighting the whole group or block, and beeping on a mismatch. It supports all the facilities offered by the Emacs 26 paren library.


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@4539 2bfe0521-f11c-4a00-b80e-6202646ff360