Standard LISP on Emacs Lisp updates
Avoid including in fasl files the code that causes a function or macro definition to be compiled because the function or macro definition has already been compiled. This saves around 6% on the size of fasl files.

However, compiled macros seem to cause problems that I am still investigating. I think this is because of the difference between the way macros work in Standard and Emacs Lisp, and converting is trickier with compiled code. I may need to keep both Standard and Emacs Lisp versions of macros! In the meantime, I will probably comment out the code that compiles macros again.

Revise GETD. This was originally an attempt to stop explicitly storing "ftype" (EXPR or MACRO, in practice), but it turns out that I need to distinguish between Emacs Lisp macros that I use to implement parts of ESL (which REDUCE doesn't expect to be macros, so I pretend they are not), and macros defined in the REDUCE source code. I also include an attempt to convert compiled Emacs Lisp macros to Standard Lisp form, but I don't think this is working correctly and it's not very elegant anyway. However, the comments are now a bit tidier!

Correct REMD to return the same value as GETD, which I missed before, although it doesn't obviously solve any problems so it probably isn't used in practice.


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@4545 2bfe0521-f11c-4a00-b80e-6202646ff360