Web site
Add a new Tutorials page and move the information about the interactive lessons from the Support page to the new Tutorials page. (I plan to add more information to the Tutorials page as time permits.) Mention the Emacs and TeXmacs interfaces on the home page, with a link to the Support page, and include the Pure project on the Projects page.


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@4640 2bfe0521-f11c-4a00-b80e-6202646ff360