GNU Emacs REDUCE IDE
In reduce-run.el: Allow the reduce-run-commands option to be customized before reduce-run has been fully loaded.  When sending REDUCE code from a file to a REDUCE process, don't switch to the process window unless this is explicitly requested (by an interactive argument). This was supposed to be the behaviour, but I think a fairly recent round of changes must have broken it. This change tidies up the code quite a lot.


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@4641 2bfe0521-f11c-4a00-b80e-6202646ff360