Alan Barnes noted (a while ago) that with "on rational,complex" certain expressions where printed in a peculiar way, namely:
a-1-i;

produces output (whether fancy is on or not)

a+(-1) -i

and

a-1/2-3i/4;

produces

       - 1      3
a + ------ - ---*i
        2       4

A complex rational number with negative imaginary part is converted to the prefix form

  (difference <realpart> <imaginarypart>)

When this appears as part of a sum 

  (plus X (difference Y Z) W)

it is printed with extra parens. It needs to be converted to

  (plus X Y (minus Z) W)

to prevent this.

In addition, a negative real part was converted to

  (quotient (minus 1) 2)

instead of 

  (minus (quotient 1 2))

which accounts for the minus sign in the numerator above.

A side effect of this change is that complex numbers may now be broken across output lines (see change in output of roots.tst file). 



git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@4653 2bfe0521-f11c-4a00-b80e-6202646ff360