REDUCE on Emacs Lisp
Replace the definition of DELETE with the code given in the Standard LISP Report. Previously, I was aliasing it to the Emacs Lisp function delete, which may act destructively. This turns out to cause serious problems. For example, it causes the Bareiss determinant code to fail. For the first time I can run "alg.tst" to completion. This takes around 1 second, which is about 20 times slower than the time shown in "alg.rlg". I don't think the results are all correct, although some of the differences from "alg.rlg" may be due to a different term ordering.


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@4699 2bfe0521-f11c-4a00-b80e-6202646ff360