REDUCE on Emacs Lisp
Implement DELETE properly.  It was supposed to call itself recursively, not call the ELisp delete function!  The "alg.tst" file finally runs correctly and there don't appear to be any differences caused by term ordering.

Implement RPLACA and RPLACD slightly more safely; they now explicitly return new cons cells.  This change is probably unnecessary but I thought it was the cause of the problem that turned out to be caused by DELETE.  Implement PROP and PLIST as in PSL and CSL as a trivial convenience.


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@4704 2bfe0521-f11c-4a00-b80e-6202646ff360