GNU Emacs REDUCE IDE
Re-implement the code to generate TAGS files for REDUCE procedure calls.  I no longer use a shell, which avoids a lot of problems and should work independently of OS.  The new implementation is synchronous, but it seems to be fast enough for this to be acceptable.  I now support tagging all REDUCE files under a directory by searching recursively.

Fix a compiler warning about `reduce-show-delim-mode' and tidy up the way it is called.


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@4709 2bfe0521-f11c-4a00-b80e-6202646ff360