GNU Emacs REDUCE IDE
Make the display of error messages by ERRORSET more robust.  It now assumes that an error message is either a string or a list, but not necessarily of strings.

Update README.txt.  The current status is that I think the lack of big integer support in Emacs Lisp is causing some tests examples in the poly package to fail.  So I need to check this and then probably implement big integers.  There is code around, but I will need to integrate whatever I use into the standard arithmetic functions.


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@4712 2bfe0521-f11c-4a00-b80e-6202646ff360