REDUCE on Emacs Lisp
Use reduce-mode in the buffer displaying REDUCE output (instead of the Lisp mode that I was using).

Re-implement LENGTH using the code in the Standard Lisp Report.  The text in the report only refers to the length of a list, but the code also applies to atoms and dotted pairs.  I had implemented LENGTH as an alias for the Elisp length function, which does not accept atoms or dotted pairs.  This broke a few examples in the int test file.

Revise READCH so that an escaped letter does not have its case changed.  This allows me to call any Elisp function from REDUCE and is more flexible that the !\ word escape I had previously implemented within COMPRESS, which I will probably phase out.  Also, I think the new behaviour is more consistent with PSL and CSL.

Replace the REDUCE function gcdn with the big integer version in the Emacs Calc package.  The former is recursive and runs out of stack on some int test examples, whereas the latter is non-recursive and works better in Emacs.

The factor test file runs correctly and the int test file runs correctly except for one example, namely int(x**7/(x**12+1),x).  This appears to involve very large numbers, which cause the Calc quotient function, which is written recursively, to run out of stack.  However, if I evaluate this integral twice, it runs on the second occasion, although the order of terms in the output is different from "int.rlg".  This is one of several quirks in Emacs REDUCE that I don't currently understand!


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@4738 2bfe0521-f11c-4a00-b80e-6202646ff360