Fluid and global declarations in ESL
According to the Standard Lisp Report, it is an error to declare a fluid variable to be global or vice versa.  However, the REDUCE (3.8) tmprint module declares the fluid variable lispsystem* to be global.  Maybe this happens elsewhere as well.  Emacs Lisp does not distinguish between fluid and global variables.  (It doesn't really have these concepts at all.)  Therefore, I now generate only a warning when a fluid variable is declared to be global or vice versa, as PSL (but not CSL) appears to do.

I have also increased the linelength used when outputting the .el files used to generate fasl files from 80 (which appears to be the default in both PSL and CSL) to 120, which should look slightly better on modern displays.  These files are peculiar to Emacs REDUCE and I originally intended them to be transient, but they turn out to be extremely useful for Lisp-level debugging!


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@4758 2bfe0521-f11c-4a00-b80e-6202646ff360