Big integer and error corrections
Make CONSTANTP and PAIRP regard a bigint as a number (i.e. a constant and not a pair).  Revise COMPRESS, EXPAND, PRIN1 and PRIN2 to use stub functions to handle bigints to facilitate the use of different bigint support to replace Calc.  (These stubs have the same names as functions in the new bigint library that I am developing.)

Remove spurious asterisks from some error messages.


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@4761 2bfe0521-f11c-4a00-b80e-6202646ff360