Emacs REDUCE
Fix a typo in "esl.el" in the previous check-in.

I want to be able to run ESL in batch mode in order to build REDUCE more conveniently.  To facilitate this I now require two more libraries explicitly in "esl.el".  (They seem to get autoloaded interactively but not in batch mode.)

Previously I introduced \ as a special escape character to force the following identifier to remain in lower case so that I can call Elisp functions from REDUCE.  I assumed that \ was not used anywhere in REDUCE, but it turns out that I used it for set difference in the sets package I wrote ages ago.  Oops!  So I have now changed the escape sequence to broken bar (U+00A6, the most obscure character on my keyboard).  I decided to keep this escape facility for now because it's a lot nicer than typing a lot of ! characters, but eventually I may drop it.  (The fact that broken bar is not ASCII may cause problems.  One has to be careful about the coding system used for files and it messed up this check-in when I tried to include it explicitly!)

In "reduce.el", append the default directory to `load-path'.  This is necessary, at least for the time being.  Change the function name `reduce' to `REDUCE' because `reduce' is an alias for `cl-reduce' in the Common Lisp emulation library, which does something completely different!  Stop using `reduce-mode' in the REDUCE buffer, at least while I get batch mode to work.  The whole user interface needs to be redesigned.


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@4763 2bfe0521-f11c-4a00-b80e-6202646ff360