Building Emacs REDUCE
In "esl.el":

When loading a Lisp library, disable the ESL mechanism to save and restore property lists when the switch DEFN is ON.  (Fixing this allows me to build the fide module.)

Tidy up FASLEND.  The main change is to detect when no Lisp output has been generated, to output a warning rather than throw an error, and not to write an empty file.  Drop the `***' prefix for messages that provide information rather than warnings.  (The REDUCE 3.8 file "misc/support.red" is effectively empty and produces no fasl output, which is not an error.)

In "reduce.el":

REDUCE is the name of a procedure in the compact module, so I cannot use it as the name of the function to start Emacs REDUCE.  Rename this to `esl-reduce'.  (The name `reduce' is already used in the Emacs CL package, which I use in ESL.)


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@4778 2bfe0521-f11c-4a00-b80e-6202646ff360