Building Emacs REDUCE
In "esl.el":

Emacs Lisp regards an identifier beginning with : as a keyword and complains if it is used in some other way, such as to name a function.  Previously, I left the ! escape in front of a : that was followed by a letter or digit, but the numeric package uses !:!:quotient as a function name, so now I leave the ! in front of a : that is followed by at least one further character, because the single-character identifier : must not be treated this way.  This allows me to build the numeric package without spurious warnings.

Add a macro called LOAD, which is provided by PSL and used in the plot package.  This is really just a place holder for now because the PSL module that plot wants to load doesn't exist in Emacs Lisp anyway, but it avoids a void function error.

Turn off warnings from the byte compiler about attempts to use the old CL library.  I don't use it in ESL (I only use the new CL library).  This allows me to build the redlog package without spurious warnings arising from the use of cl (for common logic) as a prefix.

In "remake-one-package.sh": Turn off `debug-on-error' when building the plot package, because plot is supposed to throw an error if the library fcomp is missing, which it is in Emacs.  This may no longer be necessary because the LOAD macro no longer throws an error when a library is missing, it just returns nil.

In "remake-all-packages.sh": Comment out the build of the tmprint package.  This is very Lisp-specific and would probably need some modification to build in ESL, but there would be no use for it at present.  (Perhaps I'll add fancy output to Emacs REDUCE at some later date.)


I think I can now build all REDUCE 3.8 packages except tmprint and possibly plot, so the next step is to test them all.


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@4779 2bfe0521-f11c-4a00-b80e-6202646ff360