Building Emacs REDUCE
This check-in tidies up a few loose ends.

In "esl.el": Correct the handling of identifiers beginning with : so as not to mangle the REDUCE prompt.  Improve READCH so that it handles end-of-line better.  Batch-mode input now works correctly and there are no spurious blank lines in echoed input.  This change lets me write "remake-one-package.sh" as I want to.  I have also added an attempt at an automatic check for errors to the end of "remake-all-packages.sh".


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@4780 2bfe0521-f11c-4a00-b80e-6202646ff360