Emacs REDUCE
In "esl.el": Treat !T specially so that COMPRESS does not down-case it. This seems to avoid the mangled FORTRAN output in the gentran and fide tests.  Modify ORDERP to ignore case, which is necessary since t stays lower-case when all other letters get up-cased (internally).  This seems to fix the sort-order problems in the pf and excalc tests, but not in fide (which presumably does its own sorting).

In "eslrend.red": Set symchar!* := t to get changed prompt when in symbolic mode.  (Not yet tested.)


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@4791 2bfe0521-f11c-4a00-b80e-6202646ff360