Emacs REDUCE
PUTD now respects *REDEFMSG.  Re-implement SECOND, THIRD, FOURTH, LASTCAR, NTH and PNTH as defsubsts (i.e. inline functions).  Using defalias for the first three didn't work because CADR etc. use compiler macros that are specific to the names CADR etc.  Re-implement ERROR1 to throw a silent error using a new error type, user-error-no-message.  My previous attempt at this was inelegant and didn't work.  This should suppress the myriad "***** nil" messages.  LOAD no longer loads a file if it was specified as an identifier (rather than a string) and has already been loaded.  This is consistent with the PSL definition and seems to be the expected behaviour.  Remove the "%+" continuation character output by the prettyprinter when used to generate fasl files because Emacs Lisp can't read it!  It messes up very large integers, in particular those computed at compile time to support the bernoulli function in the specfn package (in REDUCE 3.8).

Output an alert when the longer-running batch-mode scripts terminate.  Use $() instead of `` to run find in "check-one-test.sh".


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@4798 2bfe0521-f11c-4a00-b80e-6202646ff360