Minor correction to rlisp/tok.red
The first return statement in the definition of read!-comment at the end of the file was misread as a variable in the initial bootstrapping process because the code to read it correctly is in rlisp/block.red, which is read AFTER tok.red.  I have therefore added an explicit argument of nil, which seems to fix the problem.


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@4850 2bfe0521-f11c-4a00-b80e-6202646ff360