Further minor correction to alg/sub.red
Delete the conditional code in procedure subeval in alg/sub.red (not in rlisp/sub.red, as incorrectly stated in my previous check-in message) to handle the case that unwind-protect is not available.  It seems superfluous since unwind-protect is used unconditionally in several other places in REDUCE and all the Lisps on which REDUCE runs or is likely to be run provide unwind-protect.


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@4871 2bfe0521-f11c-4a00-b80e-6202646ff360