Expanding inline functions with `on defn'
When the defn switch is on, REDUCE should display the Lisp code that it would generate when the defn switch is off.  However, when REDUCE source code is input from a file with the defn switch on, only inline functions that were already defined are expanded and any inline functions defined in the file being input are ignored, whereas when the defn switch is off they are also expanded.  I think it would be more useful if inline functions defined in the file being input were also expanded when the defn switch is on.  Moreover, I need this for fasl generation to work correctly in REDUCE on Common Lisp or Emacs Lisp!

By `inline functions' I mean procedures declared as inline or smacro, or getter/setter functions declared as accessors.  These all generate a putc form, which when evaluated adds an appropriate inline property to the function identifier.  When the defn switch is on, this putc form is displayed but not evaluated.  (The identifier putc is flagged eval, so that if a call of putc were input with the defn switch on then it would be evaluated, but that doesn't help when the putc form is generated from other input.  In fact, I think it's unlikely that a call of putc would appear explicitly in a REDUCE input file.  Nevertheless, the fact that putc is flagged eval implies that putc forms should be evaluated when the defn switch is on.)

To see an example of the problem, turn on the defn switch and then input "xideal/xstorage.red".  The accessors declaration generates a number of putc forms but the accessors in the subsequent code are not expanded.  Now turn off the defn switch, load the xideal package (which defines all the inlines that it uses), turn on the defn switch and input "xideal/xstorage.red" again.  Now the accessors are all expanded.

The code that generates putc forms is in procedure formproc in "rlisp/proc.red" and in procedure expand_accessor in "rlisp/smacro.red".  I have modified this code to evaluate putc forms immediately if the defn switch is on (as well as processing the forms as before).  When the defn switch is off there should be no change from the current behaviour.


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@4968 2bfe0521-f11c-4a00-b80e-6202646ff360