Speed up the crack/v3tools test
When crack/v3tools.tst is run, procedure permu_repi in crack/v3tools.red gets called as

permu_repi '(a a a a b b b b u u u);

It essentially generates all distinct permutations of the list recursively, which is very slow and uses a lot of stack.  I have modified it to cache function values and generate only unique permutations, rather than generate all permutations and then remove duplicates.  The result is very much more efficient on the call shown above, which I hope is a typical use case.  (However, it would be much less efficient if there were no repeats in the list to be permuted.  If necessary, the strategy could be tailored to the form of input.)

I made this change in order to be able to run the v3tools test on CLISP, which seems to be quite slow and not to handle recursion very efficiently.  (I also had to re-implement the length function in Standard Lisp on Common Lisp iteratively instead of recursively so that it would run on the output of the above call in CLISP, but that's probably a worthwhile improvement anyway.)


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@4985 2bfe0521-f11c-4a00-b80e-6202646ff360