One bug and 2 'features' associated with the heuristic GCD algorithm corrected: (heugcd.red) replace the global ee by !!ee
The file heugcd.red moved from directory packages/alg to /packages/poly
where arguably it belongs and ...
(poly.red) add heugcd to create!-package poly command so heugcd code is always
loaded.  The alternative was to make heugcd into a separate package and ensure
it is loaded by ON HEUGCD by adding a simpfg property to heugcd.

(gcd.red)  avoid calling heu!-gcd from procedure gcdk when EZGCD and
HEUGCD are both ON to accord with documentation in the manual.

(ezgcdf.red) Note heu!-gcd is still called from gcdlist2 when EZGCD and
HEUGCD are both ON.  Not clear whether this is more efficient or not,
but it seems to do no harm.



git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@4995 2bfe0521-f11c-4a00-b80e-6202646ff360