Minor correction to pgauss test file
Running the file pgauss/pgauss.tst produces syntax errors and general nonsense for the last third.  The reason appears to be that the code is in symbolic mode but is being processed in algebraic mode.  I have added a symbolic declaration where I think it should be and now the code produces what look like reasonable results.


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@5006 2bfe0521-f11c-4a00-b80e-6202646ff360