Substantial revisions and corrections to elliptic function modules of the the Special Function package specfn.
In the original version the parameter m in the Jacobi functions
is the square of the modulus; in the revised version m is everywhere
the modulus; this accords with standard notation (at least in UK).
See e.g.: DF Lawden, Elliptic Functions & Applications, Springer-Verlag
(App Math Sci series #80) 1989.

Correction of several of the original rules particularly derivatives wrt to the
modulus. Some optimisations of the numerical code to avoid repeated
computations and modification of the method of evaluation for complex values
of the argument. 

Generalised shift rules have been added to remove integer multiples of K and
i*K' from arguments of the Jacobi's elliptic and Zeta functions.  This renders
most of the rules for special arguments redundant.  Miscellaneous rules have
been added where it was thought appropriate for the complete and incomplete
elliptic integrals of the first and second kind and for Jacobi's zeta function
and amplitude.

There was some original confusion between Jacobi's form of the standard
elliptic integral of the 2nd kind E(u,k) and Legendre's form D(phi,k)
where  D(am(u,k), k) = E(u, k). In particular e_function(u, k) originally
calculated D(u,k)! The e_function has been renamed as the d_function and the
e_function redefined.  Separate rules have been added for E(u,k) and D(u,k).

prifn's and fancy-prifn's have been added to produce more standard output (i.e.
less verbose).

Known problems:
In the numerical evaluation of the 4 Jacobi theta functions rounding is
a problem particularly for values of the nome near to unity (say, >0.7).
Some greater use of the descending Landen transformation is probably needed
to remedy this.

With the above mentioned exceptions, numerical evaluation of the elliptic
functions and integrals seems reasonably robust provided the modulus is real
and satisfies |k| <= 1 (even for complex arguments).  For other values of the
modulus the behaviour of the system is erratic (at best) and should not be
relied upon.  Turning ON the switch COMPLEX seems to improve matters
considerably.

Currently power series expansions of the Jacobi elliptic functions wrt to the
modulus (about k=0 in particular) is not possible  irrespective of whether
tps or taylor is used. Evaluation of the respective derivatives at k=0 produces
0/0 although all are finite-valued (the odd ones are zero).

Future Work:
Using rule-lists to invoke the numerical evaluation code  when rounded is ON
makes it difficult to control the evaluation and to avoid unnecessary repeat
calculations. Conversion of the routines to symbolic mode and interfacing in
the standard way via the properties !:rd!: and !:cr!: of jacobisn etc should
be reasonably straightforward.

Support for Weierstrassian elliptic functions and sigma functions?






git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@5062 2bfe0521-f11c-4a00-b80e-6202646ff360