REDUCE web site
Update the libraries used to the latest versions.  Move the code that
loads the JavaScript libraries required by Bootstrap into a new
include file so that they only need to be updated in one place.
Double-up backslashes in Windows filenames and replace LaTeX commands
with HTML character entities (because PHP regards backslash as an
escape).  Automate the copyright year in the page footer.  Make a few
other minor revisions.  Reformat the README file as markdown and
improve the description of how the REDUCE web site works.


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@5103 2bfe0521-f11c-4a00-b80e-6202646ff360