Some corections to the generalised shift rules capturing the periodicity and quasi-periodicity of the theta functions.
Substantial revisions to the numerical code for the theta functions


a) to use the periodicity and quasi-periodicity to reduce the argument to lie in the region bounded by the rectangle
         (0, 0), (pi/2, 0), (pi/2, -i*log(q)/2) and (0, -i*log(q)/2).

b) changes to the use of the Jacobi identity to speed up convergence when |q| is near to 1.
   For real nome q this is used when |q| >= 0.7 and otherwise for |q| >= 0.9.
   However some issues remain for complex nomes regarding the multi-valued nature of theta1 and theta2
   and also with the growth of rounding errors when its modulus is near to 0.9.

An update to sfellip.rlg to reflect the bug corrections.


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@5116 2bfe0521-f11c-4a00-b80e-6202646ff360