REDUCE web site
Add a "canonical" link element to the main template to point search
engines to the primary URL scheme and path
https://reduce-algebra.sourceforge.io rather than any of the secondary
URLs that redirect to the primary URL.

Add an empty alt attribute to the busy bee logo and delete the content
of the iframe elements on the home page to comply with the XHTML5
standard.

Delete the now redundant header block in the index template.

Delete a standard distributed Smarty PHP file from Subversion that
should never have been checked in; I don't know how that happened!


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@5126 2bfe0521-f11c-4a00-b80e-6202646ff360