REDUCE web site: add a search page.
This is experimental.  The new search page uses a Google custom search
engine using my personal Google account.  If necessary, I could add
additional administrators, or it would be easy to switch to a
different Google custom search engine using a different account.  At
some later date, it might be better to create a Google account for
REDUCE, which several developers can access, and move to using that.

It should be possible to remove advertisements given that REDUCE is a
non-profit "organisation".  But so far I haven't seen many adverts, so
this may not be necessary, although the adverts that appear presumably
depend on what you search for.  If adverts prove to be a problem then
I'll try to get them removed later.


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@5148 2bfe0521-f11c-4a00-b80e-6202646ff360