REDUCE web site: various minor improvements
Improve the appearance of the header on small displays: keep the
normal REDUCE background image if possible but shrink it as necessary
on small displays, and only switch to a simple coloured background
when absolutely necessary.

Change the alt text of the SourceForge logo to "SourceForge" instead
of "Download REDUCE" (which is the text recommended by SF).  The new
text makes more sense in the context in which I use the logo.

Indicate that the Google custom search results will include the manual
etc.

Try to ensure that the markup is valid XHTML5.  In particular, replace
the invalid search link button on the header with an anchor link
styled to look like a button.

Add "defer" to the Google translate script element so that it doesn't
cause page load to hang if it isn't available.


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@5153 2bfe0521-f11c-4a00-b80e-6202646ff360