misc/dfpart.red
Remove any existing odd, even and nonzero properties from identifiers that
are declared  generic_function.

This avoids some perhaps unexpected and incorrect simplifications of
derivatives: for example the first derivative of an odd function is assumed
odd, it is of course even. However, there is currently nothing to prevent
a user subseequently declaring a generic function to be odd or even and
obtaining incorect simplifications.

To remedy this problem substantial changes to simpdfp (or possibly simpiden)
will be needed. However, the handling of odd and even in general perhaps needs
an overhaul:

there is currently nothing to stop a user declaring something to be both odd
and even;

there is, I think, no way in algebraic mode of removing an odd or even property
from an idnetifier;

no way of, say, declaring a function of two variables to be even in the first
and odd in the second etc..



git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@5158 2bfe0521-f11c-4a00-b80e-6202646ff360