Web site: Google custom search details
I have added two XML files to Subversion in the new search directory
that specify the configuration of the Google custom search engine.
These files are not needed on the web server but they provide useful
documentation.  I have also added a brief section about the search
facility to the README file.


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@5170 2bfe0521-f11c-4a00-b80e-6202646ff360