Web site
On the External Packages page, comment out the reference to Convode,
which appears to be dead.  For some time now the web site does not
respond and I have not had any response from several attempts to email
Alain Moussiaux and one of his collaborators.  This is a pity because
I suspect it means that the Convode code is lost.  (If anyone has any
further information then please let me know.)

On the Related Projects page, add a new section at the top headed
"REDUCE for mobile devices", as suggested by Dr. Wolfgang Lindner.


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@5195 2bfe0521-f11c-4a00-b80e-6202646ff360