Overload mod to apply to both polynomials and integers
The infix operator mod has for some time had two conflicting
definitions, one in "poly/modular.red" and one in "poly/polydiv.red".
Now, mod is defined in "polydiv.red" (with higher precedence than in
"modular.red", immediately below plus) to call poly!-divide!*, which
calls evalmod if its arguments are both integers.  I have also
reinstated the Euclidean quotient operator that I originally called
div, but that clashes with other packages.  So I now call it
poly_quotient, which is not a great name but it is consistent with
pseudo_quotient.


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@5478 2bfe0521-f11c-4a00-b80e-6202646ff360