CSL: Many changes as I consolidate the "no-throw" version into the main branch. I will delete the branch directory where I had been doing initial work on it soon and that should not be used now.
OK so what is this about? The emscripten/javascript/webassembly build carries
a very high burden on code that includes C++ "catch". And that implicit
catches (and re-throws) where a function has a local instance of a class
object that has a destructor. This overhead applies regardless of whether
exceptions actually arise! It is there because (at present) webassembly does
not support exceptions, so the code does a major diversion back into JS and
then returns to wasm every time there is such a function!

In "lispthrow.h" I now have some macros that provide an abstraction over
exception handdling. It is not general, but suffices for the use required
here. For use without native exceptions I need to put use of the macro
"errexit()" just after [almost] every function call that could potentially
"throw". At present I believe I have that in around 40% of the locations that
will eventually be needed! But these ONLY matter when exceptions actually
arise, which is perhaps uncommon enough that I can survive a bit. This
risk or issue should not impact standard builds where C++ exceptions are
supported! If trying a configured with "--without-throw" it may be useful to
be aware that when a function exists with the "fake exception" flag set the
value it returns is (almost always) a special "error value", and there are
a collection of places where at least debug builds check for such values
and abort if they are seen. That should only happen if an "errexit()" is
missing, but is there to help detect when that situation arises.

Another change I am in the process of making is to remove uses of "push" and
"pop" as calls that need to be matched in favout of use of a class "Push"
that does a "push" when constructed and a pop when destroyed. It is not
quite clear that this always makes the code prettier but it is probably
safer. Well I have done not much more than 10% of that so far! But over future
checkins I hope to keep going.

This new checkin hits almost every source file. I have tested it with Windows
builds of a standard and a no-throw variant by running "testall.sh". Once
it is checked in I will have a chance to test on Linux and Macintosh. 
I *THINK* that I have seen both the regular and the without-throw versions work
but there are some delicacies and so I am not 100% certain. Testing against a
definitive checked in version will help me clear my thoughts.
But then I will want to review its impact on the embedded builld including via
emscripten and on the experimental "arithlib" and "conservative gc" work.
Each of those may be temporarily broken so those who rely on them may need
to revert to a slightly earlier checkin. As well as alerting me!


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@5550 2bfe0521-f11c-4a00-b80e-6202646ff360