sfother.red: correct a typo in the last subversion. The number!-of!-args property of StruveL was set to 1 rather than 2.
The stricter error checking of calls to functions in the specfn package in the last subversion has revealed:

1) a fairly long-established typo in specfn.tst where the call beta(1,1,1)
appeared when (presumably) ibeta(1,1,1) was intended. This has been corrected
and specfn.rlg updated.

2) two issues with the defint package. In defintb.red gamma is called with two
parameters which falls foul of the new argument checking as gamma is unary.
It is not clear yet whether a call to the binary function igamma or to the unary
function was intended (or perhaps there is a new overloading of gamma used
only in the defint package).

Similarly in definte.red, hankel2 is called with three parameters instead of
two.  This is probably due to a new overloading of Hankel2 in the defint
package to represent the Hankel transform as well as a call to the Bessel
function of the third kind, Hankel2.

dilog.red & sfbes.red:
Until these issues are defint-ly resolved, two temporary hacks have been
introduced to allow gamma to accept either 1 or 2 arguments, and hankel2
to accept either 2 or 3 arguments. This allows the unchanged defint package
to compile whilst having minimal impact of the argument count checking of
specfn calls.


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@5572 2bfe0521-f11c-4a00-b80e-6202646ff360