HTML manual navigation
The navigation within chapters in the HTML version of the REDUCE
manual was broken for chapters after the User Contributed Packages. (I
suspect it has been like this for quite a long time.) I tracked the
cause down to use of the LaTeX caption package. Removing this from
manual.tex fixed the navigation errors but broke the Excalc, Orthovec
and SPDE sections. I fixed these by adding table environments to
excalc.tex and spde.tex, and adjusting the use of the table
environment in orthovec.tex. I believe that the new version is
identical with the previous version except that the HTML navigation
now works and the vertical spacing is slightly different in the SPDE
section of the PDF version. I also fixed a misuse of \large in
rataprx.tex that caused all following text to be large in the HTML
version.


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@5612 2bfe0521-f11c-4a00-b80e-6202646ff360