Manual: re-implement syntax environment etc.
I am experimenting with using MathJax to display the maths in the
REDUCE manual, which looks good where it works, but some of the markup
causes problems. One is the use of math mode within the definitions of
the syntax and syntaxtable environments, which cases LaTeX errors
(only when generating output to use MathJax):. There is essentially no
mathematics in these environments, so I have re-implemented them not
to use math mode in their definitions. (It is OK to use math mode
within these environments, just not in their definitions!) The new
versions require a little more care over spacing and I have gone
through all their uses and revised this. I think the PDF and
conventional HTML versions are still correct.

I noticed that in the PDF version, BNF productions are indicated by a
long right arrow, whereas in the HTML version the more conventional
::= symbol is used.  I think the reason is that TeX4ht has built-in
support for \bnfprod to render as ::= and it can't be redefined.  I
have therefore renamed \bnfprod to \BNFprod, which now appears as an
arrow in both versions, as I think is intended. I also fixed a few
typos.


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@5637 2bfe0521-f11c-4a00-b80e-6202646ff360