Oops! This corrects my last commit which had the wrong message file attached.
Substantial changes to sfellip.red, sfellipi.red:
All code relating to Theta functions has been moved to a new module sftheta.red.

The code moved has been substantially restructed: firstly the second parameter
to the 4 theta functions is now the modulus tau rather than the nome q. Thus
theta1 and theta2 are now single-valued functions of tau (rather than 4-valued
functions of q related by the 4 fourth roots of unity).  The main aim is to
facilitate the future implementation of sigma functions and Weierstrassian
elliptic functions. However, the changes also make it easy to implement the 
modular transformations (tau-shifts and the Jacobi transformation) to speed up
the convergence of the Fourier series when the theta functions are evaluated
numerically.  Now no series is summed for |q| greater than about 0.2 and
convergence is extremely rapid even if the modulus of original nome is very
nearly one.

specfn.red:  sftheta added to the members of the specfn package.

Minor change to oper.tex and a more substantial one to specfn.tex to document
these changes.


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@5647 2bfe0521-f11c-4a00-b80e-6202646ff360