HTML manual formatting
When using HTML+MathJax output, macros and environments used in math
mode must be defined in math mode (on every separate web page that
needs them) in order to be available to MathJax.  This needs
conditional code since it doesn't work when not using MathJax.  The
conditional code I originally used only detected use of TeX4ht and the
rest of the code was a bit of a mess to accommodate the fact that
TeX4ht may or may not be generating MathJax output.  I have now
changed the test to detect specifically the use of MathJax output,
which allows me to make the rest of the code a lot more elegant.

I think the conditional code also caused the problem that when I built
the HTML manual without using MathJax output on Ubuntu, a spurious
mathematical image appeared at the location of the macro definitions,
where there should be no output at all.  (There was no problem if I
built the manual on Cygwin, which is why I didn't spot it for a while!
I thought the problem might be the version of TeX4ht in use, but it
doesn't appear to be that.)  I hope that handling the conditional code
better will remove this bug, but I can't easily check until I have
made this commit!

I have also noticed that building the PDF manual after having built
the HTML+MathJax version can cause a strange error, which seems to be
fixed by first deleting the .aux files, so I have modified mkpdf.sh to
do that.  It's no loss since the build requires multiple LaTeX runs
anyway.


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@5703 2bfe0521-f11c-4a00-b80e-6202646ff360