Manual formatting
The indexing of the arithmetic operators in section 3.1 was causing
spurious additional white space, which I have removed.  (I have also
introduced a \ at the end of one line, which looks redundant but is
necessary for the HTML version; TeX4ht seems to ignore white space
after } when normal LaTeX doesn't!)

The ^ symbol in the index in the HTML version was not displayed, which
I have fixed by adding another clause to the sed index post-processing
command.

I have set a maximum body width for the HTML version intended for use
on the web site (only, so far), since the manual is designed for a
printed page and looks strange with a very wide text width.  I have
also removed some redundant CSS and generally tidied the CSS up a bit.


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@5706 2bfe0521-f11c-4a00-b80e-6202646ff360