Equation references in the HTML manual using MathJax
MathJax supports \ref{} and \eqref{}, which must appear verbatim as
text within the HTML.  The \eqref macro can easily be redefined to
output itself verbatim, but \ref should only be output verbatim when
it refers to an equation.  I have therefore defined a macro \eqnref in
manual.tex, which outputs \ref{} verbatim when generating MathJax
output and calls \ref otherwise.  Uses of \ref that relate to
equations must be changed to \eqnref for this mechanism to work, which
I have done as necessary.  I have also fixed some typography.


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@5710 2bfe0521-f11c-4a00-b80e-6202646ff360