Web site
General somewhat-overdue update, including updating the Bootstrap,
jQuery and MathJax JavaScript libraries to their latest stable
released versions.  I have replaced downloading.php and windows.php
with building.php, which gives a quick general outline of building
REDUCE but defers to README.BUILDING in the repository for the
details.  It needs a little more work to make links to sections of the
HTML manual reliable and I would like to add a site map.


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@5769 2bfe0521-f11c-4a00-b80e-6202646ff360