Web site
Handle the links to the HTML REDUCE manual towards the bottom of the
Features page by using JavaScript (mostly jQuery) to look up a link
from some key text in the section title.  This avoids the links
breaking when the manual is rebuilt, since any change to the section
structure changes the links.  (There should probably also be an
external API for linking reliably to manual pages, which I'm currently
thinking about.)  I should now be able to upload the latest version of
the manual without breaking these links!  Also, update the README
file.


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@5774 2bfe0521-f11c-4a00-b80e-6202646ff360