Web site: external linking to the manual
Add the page /manual-lookup.php that is intended to be accessed with a
query string to look up HTML manual pages via their title content as I
now do internally, since the links generated by TeX4ht can change when
the manual is rebuilt.  This version uses JavaScript and the current
implementation shows a flash of the underlying manual-lookup page,
which may prove unsatisfactory in practice, but I think this code
works.


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@5775 2bfe0521-f11c-4a00-b80e-6202646ff360