Web site error reporting
Use a PHP file instead of an HTML file to report 404 errors, display
the URL that was not found and offer some advice on how to find pages.
Improve the styling and use the same styling when manual-lookup.php
reports that it has failed to load the REDUCE manual contents page.


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@5792 2bfe0521-f11c-4a00-b80e-6202646ff360