Web site map
Add a list of recently modified pages as a second tab on the site map
page.  I have not linked this page into the site navigation yet
because it seems to run rather slowly, so I need to fix that first,
which I think I can do by using caching.  I also filter out
robots.txt, which is not intended for user access.


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@5811 2bfe0521-f11c-4a00-b80e-6202646ff360