I hope that this checkin completes (at least almost) a pass of optimisation. This has involved 3 sorts of activity, these mainly being driven by a bout of profiling. (1) A significant number of small functions across Reduce have been made "inline". SOME of those had been ones that were identifiable as hot-spots, while others are changes just because the nice abstraction that is "symbolic procedure component2 u; cadr u;" that makes the data structure componnent more abstract ought not to add run-time cost. To make this work I needed to upgrade the code in rlisp/smacro.red so that inline functions should now evaluate just the same things and in just the same order as regular functions. (2) Some parts of the Reduce code had been written either for clarity or convenience or merely without understanding of performance consequences. That only matters when the functions involved end up really heavily used, but it turned out that in some cases that was the situation. One particular issue was code that conflicted with the CSL model for profiling code to identify hot-spots and mapping those functions into C++. That could bite both in cases where people had introduced redefinitions of functions from the Reduce core and when a function set up as an autoload entrypoint was heavily used. Both of those caused CSL to avoid mapping things into C++ because of risk of confusion. I updated CSL so it could code more gracefully with multiply defined functions. I continue to believe that the presence of those is unsatisfactory from at the very least a code transparancy perspective, and it would be a blockage to global static analysis, so I want a long term sweep to remove them all! (3) The existing code made expecially heavy use of some functions where the CSL implementation was coded to be clean and general because it had not expected intensive use. That especially applied to explode, compress and intern where the code appeared to have been written based on a believe that they were cheap operations, which is a questionable matter. I both re-worked parts of CSL to have special cheap paths through the cases I believe that Reduce was using most and I adjusted the Reduce code to cache some of the results, use different and less expensive Lisp functions and in general attempt to clean things up. As with the multiply defined functions I concentrated on procedures that were heavily used and where changes where most liable to impact overall speed, and in a very few places I re-coded some "hot" functions in ways that may be less elegant but that avoid extraneous consing or other work. The Lisp function id2int, which is not part of Standard Lisp, gets used to provide hash values, so I re-worked CSL so that it returns something at least vauguely suitable for that rather then the version where I had previously guessed what might be a sensible definition. While doing this I have made (small) improvements to the CSL compiler and to the build process. If I use revision 5750 as a baseline and quote timings collected on a tolerably fast Windows-10 machine the impact of all of this for Reduce as a whole is as follows: If I run "scripts/testall.sh --noregressions" then I see    5750 CSL   100.716    new CSL     83.971    (about a 16.6% improvment!)    5750 PSL   146.950    new PSL    136.817    (about 7.25%) The greater speedup for CSL is because of a mix of 5750 defeating the profiling and optimisation into C++ and of the new changes within CSL.
In a rather few cases the performance changes were even more dramatic. With
5750 the two packages where CSL showed up especially slow as against PSL
were pasf and ibalp. These were not chosen as parts of redlog but as the
test cases where the PSL:CSL speed ratio was most extreme. For plenty of
other cases the two systems run within say 80%-125% as the speed ratio and
that seems reasonable!
                  pasf           ibalp
               CSL   PSL       CSL   PSL
r5750         7.252 0.766     3.860 1.165
new           2.032 0.781     1.453 1.219
where here I believe the apparent slight slowdown for PSL is within the
variation that test runs tend to show, and other parts of the redlog
sub-family of tests show speedups (around 5%) with this set of updated.
The rather substantial speedups for CSL represent cases where the coding
in some packages reacted especially badly with CSL internals (or perhaps
they had been developed and carefully tuned for PSL), and the improvement
comes from changes in both CSL and Reduce itself.
I note that CSL still falls behind PSL on a range of the test cases,
and I do not yet fully understand why. PSL is slower on some other cases
and discussions with Rainer mean that at least a major part of that is
understood and may get addressed after some debugging.

A part of the problem within CSL was that the function "resource-limit", used
to set what was intended as a long-stop time limit for profiling had some
issues. While that could have impacted most of the profiling it showed up
especially sharply with ibalp.tst first because that test started with code
that allocated and recycled a great deal of memory, triggering a number of
garbage collections that could be significant in small amounts of memory. And
then the parts of the tests with that feature were followed by calls to
a different set of functions - also very important overall. But if
resource-limit (incorrectly) truncated the run that second set if functions
failed to get identified as worth optimising into C++. For many other
packages it happened that even the initial part of the test script exercised
a fairly fully representative fraction of the code. This issue was certainly
several years old.

I hope that the speedups that this all gives will be seen as a positive
move.



git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@5828 2bfe0521-f11c-4a00-b80e-6202646ff360