for-emscripten/Makefile
Update to reflect recent updates to the main CSL sources, following
advice from Arthur.  Replace references to bytes.o with bytes1.o etc.
and delete references to bytes.o where there was already a similar
reference to bytes1.o.  This allows me to progress a bit further with
an Emscripten build of REDUCE, so I think these updates are OK.


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@5830 2bfe0521-f11c-4a00-b80e-6202646ff360